+++ docsrc/source/changelog.rst
@@ -15,8 +15,10 @@ v0.28.0
 - Added datasets.info()
 - Added df.apply()
 - Added additional functionality for sampling etc.
-- Fixed documentation to better suit
+- Fixed documentation for Datasets API
 - Add new monitoring health test for chunk data structure
+- Add fix for csv reading for _chunk_ to be parsed as actual Python objects
+and not strings
 
 v0.27.0
 --------